requirements.txt
fastapi==0.61.1
fastapi-utils==0.2.1
uvicorn==0.12.1
rich==9.13.0
python-dotenv==0.14.0
gunicorn==20.0.4
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.1.48
Jinja2==3.0.1